requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
mako==1.3.10
markupsafe==3.0.2
passlib==1.7.4
psycopg==3.2.6
psycopg2==2.9.10
pydantic==2.11.3
pydantic-core==2.33.1
pyjwt==2.10.1
python-multipart==0.0.20
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.1
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.0
